functions/Backend/firebase_config/firebase_firestore.py
import firebase_admin
from firebase_admin import credentials, firestore

# Inicializar Firebase
cred = credentials.Certificate('path/to/serviceAccountKey.json')
firebase_admin.initialize_app(cred)

# Conectar ao Firestore
db = firestore.client()

def get_firestore_client():
    return db
