functions/Backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.7
Django==5.0.2
dnspython==2.7.0
et_xmlfile==2.0.0
Faker==27.0.0
filelock==3.12.4
Flask==3.0.2
Flask-Cors==5.0.0
greenlet==3.0.0
gunicorn==23.0.0
h11==0.14.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.20.1
marshmallow-dataclass==8.5.3
mypy-extensions==1.0.0
numpy==2.1.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
pillow==11.0.0
platformdirs==3.11.0
pycparser==2.22
pymongo==4.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
robotframework==7.0.1
robotframework-faker==5.0.0
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.3.0
selenium==4.23.1
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.26.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.24.5
waitress==3.0.2
websocket-client==1.8.0
Werkzeug==3.0.1
wheel==0.41.2
wrapt==1.16.0
wsproto==1.2.0

